# interop/service.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "interop/service.proto",
    "version": "version not set"
  },
  "schemes": [
    "http",
    "https"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/v1/interop/mission": {
      "get": {
        "summary": "Get active mission",
        "operationId": "GetActiveMission",
        "responses": {
          "200": {
            "description": "",
            "schema": {
              "$ref": "#/definitions/interopMission"
            }
          }
        },
        "tags": [
          "Interop"
        ]
      }
    },
    "/v1/interop/mission/status": {
      "get": {
        "operationId": "Status",
        "responses": {
          "200": {
            "description": "",
            "schema": {
              "$ref": "#/definitions/interopMissionReportStatus"
            }
          }
        },
        "tags": [
          "MissionReporting"
        ]
      }
    },
    "/v1/interop/server": {
      "get": {
        "operationId": "StartServer",
        "responses": {
          "200": {
            "description": "",
            "schema": {
              "$ref": "#/definitions/protobufEmpty"
            }
          }
        },
        "tags": [
          "Interop"
        ]
      }
    },
    "/v1/missionedit/mission/{mission_name}": {
      "get": {
        "operationId": "GetMission",
        "responses": {
          "200": {
            "description": "",
            "schema": {
              "$ref": "#/definitions/interopCompleteMission"
            }
          }
        },
        "parameters": [
          {
            "name": "mission_name",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "tags": [
          "MissionEdit"
        ]
      },
      "post": {
        "operationId": "PostMission",
        "responses": {
          "200": {
            "description": "",
            "schema": {
              "$ref": "#/definitions/protobufEmpty"
            }
          }
        },
        "parameters": [
          {
            "name": "mission_name",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/interopCompleteMission"
            }
          }
        ],
        "tags": [
          "MissionEdit"
        ]
      }
    },
    "/v1/missionedit/mission_names": {
      "get": {
        "operationId": "GetMissionList",
        "responses": {
          "200": {
            "description": "",
            "schema": {
              "$ref": "#/definitions/interopMissionNameList"
            }
          }
        },
        "tags": [
          "MissionEdit"
        ]
      }
    }
  },
  "definitions": {
    "interopCompleteMission": {
      "type": "object",
      "properties": {
        "mission_name": {
          "type": "string"
        },
        "mission": {
          "$ref": "#/definitions/interopMission"
        },
        "obstacles": {
          "$ref": "#/definitions/interopObstacles"
        },
        "sitlconfig": {
          "$ref": "#/definitions/interopSitlConfig"
        },
        "targets": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/interopTarget"
          }
        }
      }
    },
    "interopDistanceStatus": {
      "type": "object",
      "properties": {
        "closest_dist": {
          "type": "number",
          "format": "float"
        },
        "current_dist": {
          "type": "number",
          "format": "float"
        }
      }
    },
    "interopFlyZone": {
      "type": "object",
      "properties": {
        "altitude_msl_min": {
          "type": "number",
          "format": "float"
        },
        "altitude_msl_max": {
          "type": "number",
          "format": "float"
        },
        "boundary_pts": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/interopWaypoint"
          }
        }
      }
    },
    "interopGPS": {
      "type": "object",
      "properties": {
        "latitude": {
          "type": "number",
          "format": "float"
        },
        "longitude": {
          "type": "number",
          "format": "float"
        }
      }
    },
    "interopMission": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer",
          "format": "int64"
        },
        "active": {
          "type": "boolean",
          "format": "boolean"
        },
        "air_drop_pos": {
          "$ref": "#/definitions/interopGPS"
        },
        "home_pos": {
          "$ref": "#/definitions/interopGPS"
        },
        "off_axis_target_pos": {
          "$ref": "#/definitions/interopGPS"
        },
        "emergent_last_known_pos": {
          "$ref": "#/definitions/interopGPS"
        },
        "fly_zones": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/interopFlyZone"
          }
        },
        "mission_waypoints": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/interopWaypoint"
          }
        },
        "search_grid_points": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/interopWaypoint"
          }
        }
      }
    },
    "interopMissionNameList": {
      "type": "object",
      "properties": {
        "mission_name_list": {
          "type": "array",
          "items": {
            "type": "string"
          }
        }
      }
    },
    "interopMissionReportStatus": {
      "type": "object",
      "properties": {
        "mission_waypoints": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/interopDistanceStatus"
          }
        },
        "airdrop_pos": {
          "$ref": "#/definitions/interopDistanceStatus"
        },
        "emergent_pos": {
          "$ref": "#/definitions/interopDistanceStatus"
        },
        "stationary_obstacles": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/interopDistanceStatus"
          }
        },
        "moving_obstacles": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/interopDistanceStatus"
          }
        }
      }
    },
    "interopMovingObstacle": {
      "type": "object",
      "properties": {
        "latitude": {
          "type": "number",
          "format": "float"
        },
        "longitude": {
          "type": "number",
          "format": "float"
        },
        "altitude_msl": {
          "type": "number",
          "format": "float"
        },
        "sphere_radius": {
          "type": "number",
          "format": "float"
        }
      }
    },
    "interopObstacles": {
      "type": "object",
      "properties": {
        "moving_obstacles": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/interopMovingObstacle"
          }
        },
        "stationary_obstacles": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/interopStationaryObstacle"
          }
        }
      }
    },
    "interopSitlConfig": {
      "type": "object",
      "properties": {
        "takeoff_direction": {
          "type": "integer",
          "format": "int32"
        }
      }
    },
    "interopStationaryObstacle": {
      "type": "object",
      "properties": {
        "latitude": {
          "type": "number",
          "format": "float"
        },
        "longitude": {
          "type": "number",
          "format": "float"
        },
        "cylinder_height": {
          "type": "number",
          "format": "float"
        },
        "cylinder_radius": {
          "type": "number",
          "format": "float"
        }
      }
    },
    "interopTarget": {
      "type": "object",
      "properties": {
        "type": {
          "type": "string"
        },
        "latitude": {
          "type": "number",
          "format": "float"
        },
        "longitude": {
          "type": "number",
          "format": "float"
        },
        "shape": {
          "type": "string"
        },
        "alphanumeric": {
          "type": "string"
        },
        "alpanumeric_color": {
          "type": "string"
        },
        "background_color": {
          "type": "string"
        },
        "description": {
          "type": "string"
        },
        "orientation": {
          "type": "string"
        }
      }
    },
    "interopWaypoint": {
      "type": "object",
      "properties": {
        "order": {
          "type": "integer",
          "format": "int64"
        },
        "latitude": {
          "type": "number",
          "format": "float"
        },
        "longitude": {
          "type": "number",
          "format": "float"
        },
        "altitude_msl": {
          "type": "number",
          "format": "float"
        }
      }
    },
    "protobufEmpty": {
      "type": "object",
      "description": "service Foo {\n      rpc Bar(google.protobuf.Empty) returns (google.protobuf.Empty);\n    }\n\nThe JSON representation for `Empty` is empty JSON object `{}`.",
      "title": "A generic empty message that you can re-use to avoid defining duplicated\nempty messages in your APIs. A typical example is to use it as the request\nor the response type of an API method. For instance:"
    }
  }
}